dev-plotting/requirements.txt
botocore==1.23.41
cloudinary==1.28.1
Flask==2.0.2
Flask_Inputs==0.3.0
matplotlib==3.5.1
MetPy==1.1.0
nexradaws==1.1
numpy==1.22.1
pika==1.2.0
pydantic==1.9.0
python-decouple==3.6
python-dotenv==0.19.2
jsonschema==4.4.0
aioflask==0.4.0